Fix match determination
I noticed in the interactive site that many of the results that showed up only on the fuse side curiously had a score of exactly 0.6, the default threshold. Looking back at the fuse source, it seems like it's a match if it's below OR equal to the threshold. This change fixes it to match the fuse way.