Fix issue with matching at the end
Text that _ends_ with a partial match wasn't working correctly, ex a pattern of "abc" fails to match at "ab" with an edit distance of 1, even though it's just one insertion away. This is the reverse of heyimalex/bitap#1, which makes sense because we reverse the text and pattern before searching. This commit mitigates this issue, and anecdotally it seems like the output from fuse and wafu are now totally equivalent! Which is to say I _think_ this change fixed it, but I need to add more tests before being sure.