Create faster unsafe version
To avoid needing to deserialize fields/options on every call, we just create a Searcher and return a pointer. This is not super safe from the js world because callers need to remember to free or else face a memory leak, which isn't something I want to force right now. However, it makes my perf numbers look better in the demo so I'm implementing it anyway.