Use char vec instead of str
This is a quick find and replace to check whether using Vec<char> is any faster than using &str in Fields. And it looks like it pretty significantly speeds up wafu in chrome.

There were real thoughtful motivations for this, but I tbh can't remember what they were. Anyway, results say that changing to char in a more thoughtful way may be worthwhile.